Test PHP 8.3 language features
