+++ phpunit.xml.dist
@@ -33,22 +33,24 @@
       <exclude>./tests/PHP80</exclude>
       <exclude>./tests/PHP81</exclude>
       <exclude>./tests/PHP82</exclude>
-      <exclude>./tests/PHP83</exclude>
+      <exclude>./tests/Unit/PHP81</exclude>
+      <exclude>./tests/Unit/PHP83</exclude>
     </testsuite>
     <testsuite name="Mockery Test Suite PHP80">
       <directory phpVersion="8.0.0-dev" phpVersionOperator="&gt;=">./tests</directory>
       <exclude>./tests/PHP81</exclude>
       <exclude>./tests/PHP82</exclude>
-      <exclude>./tests/PHP83</exclude>
+      <exclude>./tests/Unit/PHP81</exclude>
+      <exclude>./tests/Unit/PHP83</exclude>
     </testsuite>
     <testsuite name="Mockery Test Suite PHP81">
       <directory phpVersion="8.1.0-dev" phpVersionOperator="&gt;=">./tests</directory>
       <exclude>./tests/PHP82</exclude>
-      <exclude>./tests/PHP83</exclude>
+      <exclude>./tests/Unit/PHP83</exclude>
     </testsuite>
     <testsuite name="Mockery Test Suite PHP82">
       <directory phpVersion="8.2.0-dev" phpVersionOperator="&gt;=">./tests</directory>
-      <exclude>./tests/PHP83</exclude>
+      <exclude>./tests/Unit/PHP83</exclude>
     </testsuite>
     <testsuite name="Mockery Test Suite PHP83">
       <directory phpVersion="8.3.0-dev" phpVersionOperator="&gt;=">./tests</directory>
