+++ tests/Fixture/PHP83/ClassName.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Fixture\PHP83;
+
+class ClassName {
+    public const CONSTANT = 42;
+}
