+++ tests/Fixture/PHP83/Classes.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Fixture\PHP83;
+
+class Classes implements Interfaces {
+    use Traits;
+}
