+++ tests/Fixture/PHP83/Enums.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Fixture\PHP83;
+
+enum Enums {
+    const string FOO = "bar";
+}
