+++ tests/Fixture/PHP83/Interfaces.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Fixture\PHP83;
+
+interface Interfaces {
+    const string BAR = Enums::FOO;
+
+    public function foo(): string;
+}
