+++ tests/Fixture/PHP83/Traits.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Fixture\PHP83;
+
+trait Traits {
+    const string BAR = Enums::FOO;
+
+    public function foo(): string
+    {
+        return Interfaces::BAR;
+    }
+}
